moves favourites_screen to ui/favourites
