+++ lib/services/routing_service.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:lingo_jam/ui/screens/favourites_screen.dart';
+import 'package:lingo_jam/ui/favourites/favourites_screen.dart';
 import 'package:lingo_jam/ui/stations/stations_screen.dart';
 import 'package:lingo_jam/ui/screens/test_screen.dart';
